requirements.txt
adal==1.2.7
algoliasearch==1.20.0
amqp==2.6.1
appdirs==1.4.4
appnope==0.1.0
asgiref==3.3.4
attrs==20.2.0
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.27
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.52
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
backcall==0.2.0
billiard==3.6.3.0
boto3==1.14.60
botocore==1.17.60
cached-property==1.5.1
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.6.0
Django==3.2.4
django-anymail==8.0
django-crispy-forms==1.9.2
django-dirtyfields==1.4
django-discover-runner==1.0
django-environ==0.4.5
django-impersonate==1.5.1
django-js-reverse==0.9.1
django-ratelimit==3.0.1
django-redis-cache==2.1.1
django-role-permissions==3.1.0
django-storages==1.10
django-tinymce==3.0.2
djangorestframework==3.12.4
docraptor==2.0.0
docspring==1.3.0
docutils==0.15.2
hellosign-python-sdk==4.0.0
idna==2.10
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.17.2
jmespath==0.10.0
kombu==4.6.11
lxml==4.5.2
msrest==0.6.21
msrestazure==0.6.4
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.9
pickleshare==0.7.5
Pillow==7.2.0
postmarker==0.15.0
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycountry==20.7.3
pycparser==2.20
Pygments==2.7.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
sendgrid==6.4.6
sentry-sdk==0.17.4
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
traitlets==4.3.3
twilio==6.45.1
typeform==1.1.0
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
zeep==3.4.0
zipp==3.1.0
